chore: extract query-client
